@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :line do |line|
     line.line_text { Faker::Lorem.sentence }
     line.count 0
