@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :tweet do
     after(:create) do |tweet|
       3.times.each do
